A walk keeps its order while the day moves under it
The store landed in the commit before this one. This is the assembler using it,
and the wiring that makes it real: the cursor carries the walk's id, a first
page freezes what it assembled, and later pages take their sequence from that
freeze instead of from a ranking rebuilt under the reader.

WHAT A READER GETS. The rows they started with, in the order they started in,
each exactly once — over a day where work arrives AND the ranking moves, which
is the pair the offset pager cannot survive and says so in its own contract.

WHAT THEY ARE TOLD RATHER THAN SHOWN. `walk.changed_since_snapshot` is how many
of their rows have gone: answered, deleted, or no longer visible. Those leave
immediately, because a frozen headline over work somebody can no longer see
would be steadier and false. `walk.new_available` is what has arrived behind
them, which waits for a refresh — absent on a first page, where the question has
no meaning yet.

A WALK THAT CANNOT BE RESUMED IS REFUSED, never quietly restarted. Serving a
fresh page under a resumed token would hand the reader rows they have already
seen, in a new order, with nothing saying the walk they were on had ended.

Unwired, nothing changes: an installation with no store pages exactly the way
this queue always did, and a cursor minted before this shipped still resumes.
That is the rollout lever and the honest absence, not a degraded mode.

THE ACCEPTANCE TEST FOUND A REAL BUG, and only after its fixture got harder.
rankAllFor re-sorted every page before rendering it — which is right for a fresh
cut and undoes a frozen walk two lines after its own pager got the order right.
The first fixture used tasks with monotonic deadlines, so the frozen order and a
re-ranked order were identical and the mutation survived. With the deadlines
reversed the test failed immediately, on the implementation rather than on
itself. Sorting and rendering are separate functions now: renderInOrder stamps
the wire fields onto rows already in the order they are to be drawn.

worklist.go crossed the 500-line cap, so what the day could NOT see — the lane
scan bounds and the withheld-lane vocabulary — moved to unseen.go. One concept:
a lane falls short by reading to its bound or by not answering at all, and a
queue must never read as clear over either.
